HADOOP-13230 factor out base class for cost tests and split suite by category
